+++ src/Models/NewsalertRecipientsModel.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2017 Heimrich & Hannot GmbH
+ *
+ * @author  Thomas Körner <t.koerner@heimrich-hannot.de>
+ * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ */
+
+namespace HeimrichHannot\ContaoNewsAlertBundle\Models;
+
+class NewsalertRecipientsModel extends \Contao\Model
+{
+    protected static $strTable = 'tl_newsalert_recipients';
+
+}