+++ src/Resources/contao/dca/tl_news.php
@@ -35,8 +35,8 @@ $dc['subpalettes']['newsalert_activate'] = 'newsalert_send';
 $fields = [
     'newsalert_activate' => [
         'label'     => [
-            $translator->trans('hh.newsalert.tl_news.field.newsalert_activate'),
-            $translator->trans('hh.newsalert.tl_news.field.newsalert_activate.desc')
+            $translator->trans('hh.newsalert.tl_news.newsalert_activate.name'),
+            $translator->trans('hh.newsalert.tl_news.newsalert_activate.desc')
         ],
         'inputType' => 'checkbox',
         'exclude'   => true,
@@ -45,8 +45,18 @@ $fields = [
     ],
     'newsalert_send' => [
         'label'     => [
-            $translator->trans('hh.newsalert.tl_news.field.newsalert_send'),
-            $translator->trans('hh.newsalert.tl_news.field.newsalert_send.desc')
+            $translator->trans('hh.newsalert.tl_news.newsalert_send.name'),
+            $translator->trans('hh.newsalert.tl_news.newsalert_send.desc')
+        ],
+        'inputType' => 'checkbox',
+        'exclude'   => true,
+        'sql'       => "int(1) NOT NULL default 0",
+        'eval'      => ['tl_class' => 'w50'],
+    ],
+    'newsalert_sendtest' => [
+        'label'     => [
+            $translator->trans('hh.newsalert.tl_news.newsalert_send.name'),
+            $translator->trans('hh.newsalert.tl_news.newsalert_send.desc')
         ],
         'inputType' => 'checkbox',
         'exclude'   => true,