+++ src/Resources/contao/languages/de/tl_newsalert_recipients.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2017 Heimrich & Hannot GmbH
+ *
+ * @author  Thomas Körner <t.koerner@heimrich-hannot.de>
+ * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ */
+
+$arrLang = &$GLOBALS['TL_LANG']['tl_newsalert_recipients'];
+
+$arrLang['abonnement_legend'] = "Newsalert Abonnement";